requirements.txt
pypdf
langchain
chromadb-client 
langchain_community
pytest
boto3